managed-agentcore/requirements.txt
strands-agents
strands-agents-tools
boto3
python-dotenv
bedrock-agentcore<=0.1.5
